Move routes folder under app
