+++ app/routes/api/v1/favorites.js
