+++ app/routes/api/v1/index.js
@@ -1,7 +1,7 @@
 var express = require('express')
 var router = express.Router()
 const environment = process.env.NODE_ENV || 'development'
-const configuration = require('../../../knexfile')[environment]
+const configuration = require('../../../../knexfile')[environment]
 const database = require('knex')(configuration)
 const fetch = require('node-fetch')
